requirements.txt
# Dependencies for AgentPay Simulation

# Core dependencies
dedalus-labs>=0.1.0
agentpay-sdk>=1.0.0  # Install from ../AgentPay-SDK
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# Web & API
requests>=2.31.0
aiohttp>=3.8.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.0.280

# Jupyter (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
